issue 161 (externalizing Strings)
Some of these are hard to do because the getString method is not readily available.
There are other places where the code needs to be refactored because just
externalizing Strings isn't enough. For example "5 minutes ago" in English
translates to "hace 5 minutos" in Spanish. So there is still work to be done.